backend/requirements.txt
fastapi==0.111.2
uvicorn[standard]==0.23.2
pandas==2.1.1
numpy==1.26.5
scikit-learn==1.3.1
rank-bm25==0.2.3
gensim==4.3.7
sentence-transformers==2.2.2
PyMuPDF==1.23.1
tensorflow==2.13.0
python-dotenv
pypdf
python-multipart
openai
